Create 5 round function and display results
I added every possible outcome to the single round function. I then
added a function that would iterate the single round function five
times and return the result of each round. Also added a function that
displays the winner at the end of the 5 rounds based on who's score is
higher than the other.