Add iteration decrement
Changed the file name from javascript.js to script.js to keep up
with best practices. Changed index.html to reflect. Added an additional
conditional statement to the game() function to account for an invalid
prompt input that I would like to result in a iteration decrement to
ensure the user plays the full 5 rounds. I also simplified the values
of the string outputs in playRound(), considering those are only used
as arguements in the game() function.