Build script to walk user through start prompts
Deleted html body as I will be building the UI using DOM manipulation,
including the start screen. Built the starting pop-up, asking the user
if they would like to move forward with a game. If no, it redirects to
my RPS github repo. If yes, the starting pop-up will be removed and the
rules pop-up will populate, prompting the user to start the game.