Add function to populate gameplay window
Added a function that would remove the rules-popup and populate the
gameplay window. Included is a round counter, computer and player score
displays, and weapons choices. A function was written to remove the
unselected weapons choices, leaving the selected weapons choice on the
screen. Event listeners were attached to the weapons buttons to
initiate a game round played with the corresponding weapon chosen.