Complete flexbox config of all containers
The flexbox configuration of all created DIVs is complete. What is not
visible at this point in the code is the message display box located
above the weapons buttons and the computer weapon choice display area
above the score container. When the round is played, both of these areas
will populate with content.